refactor(views): extract view orchestration with use-case names
